app/requirements.txt
python-dotenv==0.19.2
psycopg2-binary==2.9.2
confluent-kafka==1.7.0
gensim==4.1.2
nltk==3.6.5
pyLDAvis==3.3.1
textblob==0.17.1